% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callapi.R
\name{apicheckupdates}
\alias{apicheckupdates}
\title{Check if a file has been updates since it was copied}
\usage{
apicheckupdates(mrc.hash, local.hash, file)
}
\arguments{
\item{mrc.hash}{character. Hash of the most recent commit}

\item{local.hash}{character. Hash from the copied file, corresponding to the most recent commit when it was copied}

\item{file}{character. The full path of the the file to check}
}
\description{
Check if a file has been updates since it was copied
}
