% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callapi.R
\name{apicommitfile}
\alias{apicommitfile}
\title{Commit a file to a forked repository}
\usage{
apicommitfile(forkname, filename, contents, message)
}
\arguments{
\item{forkname}{character Name of the fork to make the commit to}

\item{filename}{character Name of the file to commit or change}

\item{contents}{character Contents of the new file}

\item{message}{character Commit message, include your name}
}
\description{
Commit a file to a forked repository
}
