% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callapi.R
\name{apideletefork}
\alias{apideletefork}
\title{Permanently delete a fork
Use for cleaning up forks made after files have been committed}
\usage{
apideletefork(forkname)
}
\arguments{
\item{forkname}{character. Name of the fork to delete}
}
\description{
Permanently delete a fork
Use for cleaning up forks made after files have been committed
}
