% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callapi.R
\name{apigetscriptcontent}
\alias{apigetscriptcontent}
\title{Get the content of a single scripts}
\usage{
apigetscriptcontent(mrc.hash, file.name)
}
\arguments{
\item{mrc.hash}{character. Hash of the most recent commit}

\item{file.name}{character. The full path of the file to get contents of}
}
\description{
Get the content of a single scripts
}
