% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callapi.R
\name{apigetscriptnames}
\alias{apigetscriptnames}
\title{Get a list of scripts from the BB code library through the api}
\usage{
apigetscriptnames(mrc.hash, language = "all")
}
\arguments{
\item{mrc.hash}{character. Hash of the most recent commit}

\item{language}{character. Default = all. List scripts of a certain folder or all scripts}
}
\description{
Get a list of scripts from the BB code library through the api
}
