% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callapi.R
\name{apimakefork}
\alias{apimakefork}
\title{Make a new fork of the code library}
\usage{
apimakefork(forkname)
}
\arguments{
\item{forkname}{character. The name to give the new fork}
}
\description{
Make a new fork of the code library
}
