% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callapi.R
\name{apimakepullrequest}
\alias{apimakepullrequest}
\title{Make a pull request through the API}
\usage{
apimakepullrequest(forkname, message)
}
\arguments{
\item{forkname}{character. The name of the fork to make the request from}

\item{message}{character. A message explaining what the pull request is for which goes as the title of the pull request}
}
\description{
Make a pull request from a fork to the main code library repository
}
