% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codelibrary.R
\name{check_for_updates}
\alias{check_for_updates}
\title{Check for updates on a file}
\usage{
check_for_updates(from, proj_path = ".")
}
\arguments{
\item{from}{character. file name to check}

\item{proj_path}{character. Default = current working directory. path to tidyproject}
}
\description{
Check if a locally copied script has since been updated in the BB code library by using hashes to see the diff
}
