% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codelibrary.R
\name{info_scripts}
\alias{info_scripts}
\title{List information about scripts}
\usage{
info_scripts(files, fields = c("Description"), viewer = TRUE,
  silent = FALSE)
}
\arguments{
\item{files}{vector string of file names/paths}

\item{fields}{vector string of field tags to display}

\item{viewer}{logical indicating if Rstudio viewer should be used. Default = TRUE}

\item{silent}{logical run in quiet mode (default=FALSE)}
}
\description{
Displays information about selected scripts
}
