% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codelibrary.R
\name{preview_script}
\alias{preview_script}
\title{Preview  file from the BB code library}
\usage{
preview_script(name)
}
\arguments{
\item{name}{character indicating script in the code library to preview}
}
\description{
Preview  file from the BB code library
}
