% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codelibrary.R
\name{submit_script}
\alias{submit_script}
\title{Submit a file via pull request}
\usage{
submit_script(from, message, language = "R", overwrite = FALSE,
  proj_path = ".", forkname = "qcpuserfork")
}
\arguments{
\item{from}{character. file name or path of file to submit}

\item{message}{character. The message to make the pull request with. Include name of user}

\item{overwrite}{logical. Overwrite the file if it exists in code library already?}

\item{proj_path}{character. Default = current working directory. path to tidyproject}

\item{forkname}{character. Default = qcpuserfork. Name of forked repository. Current time is added to name to prevent duplicates}
}
\description{
Makes a fork of the BB code library with the QCP user account, then commits the file to that fork, then makes a pull request back to the BB code library
Pull request is made with the message
}
